integration: move encryption assertions into test
